chore: unified file index js, rm keyword
