+++ .github/packager/js/index.js
@@ -1,4 +1,33 @@
-// Main entry point for the pactus-grpc package
-// Re-exports all modules from the pactus_grpc namespace
+const modules = {};
 
-module.exports = require('./pactus_grpc'); 
+try {
+  const fs = require('fs');
+  const path = require('path');
+  
+  fs.readdirSync(__dirname)
+    .filter(file => file.endsWith('.js') && file !== 'index.js')
+    .forEach(file => {
+      const moduleName = path.basename(file, '.js');
+      modules[moduleName] = require(`./${file}`);
+    });
+  
+  fs.readdirSync(__dirname)
+    .filter(dir => {
+      const dirPath = path.join(__dirname, dir);
+      return fs.existsSync(dirPath) && fs.statSync(dirPath).isDirectory();
+    })
+    .forEach(dir => {
+      modules[dir] = {};
+      fs.readdirSync(path.join(__dirname, dir))
+        .filter(file => file.endsWith('.js'))
+        .forEach(file => {
+          const moduleName = path.basename(file, '.js');
+          modules[dir][moduleName] = require(`./${dir}/${file}`);
+        });
+    });
+} catch (err) {
+  console.error('Error loading gRPC modules:', err);
+}
+
+module.exports = modules;
+
