+++ .github/packager/js/package.json
@@ -20,8 +20,7 @@
   "keywords": [
     "pactus",
     "grpc",
-    "blockchain",
-    "cryptocurrency"
+    "blockchain"
   ],
   "author": "Pactus Developers",
   "license": "MIT",
