+++ .gitignore
@@ -14,3 +14,4 @@ todo
 # VIM artifacts
 .swp
 .*.sw*
+packages/
